# PERF: Improve merge performance